2 seat map with selected effect
