+++ frontend/src/pages/EventDetails/EventDetails.tsx
@@ -8,7 +8,6 @@ import { Button } from 'react-bootstrap';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Avatar from '@mui/material/Avatar';
-import { SeatMap } from './SeatMap';
 
 interface TabPanelProps {
     children?: React.ReactNode;
@@ -169,7 +168,6 @@ export const EventDetails: React.FC = (): JSX.Element => {
                     </Grid>
                 </CustomTabPanel>
                 <CustomTabPanel value={value} index={1}>
-                <SeatMap/>
                 </CustomTabPanel>
                 <CustomTabPanel value={value} index={2}>
                 Item Three
