--- frontend/src/pages/Venue/MySVGComponent.css
@@ -1,3 +1,0 @@
-.selected {
-  outline: 2px solid red;
-}
