+++ frontend/src/pages/Venue/Venue.tsx
@@ -1,20 +1,21 @@
 import React, { useEffect, useState } from 'react';
 import { Navigate } from 'react-router-dom';
 import { NavbarLoggedIn, NavbarNotLoggedIn } from '../../Navbar';
-import SeatMap from './seatMap';
+import { SGStad } from '../utility/SeatMap'
+
 
 export const Venue = () => {
     useEffect(() => {
     }, []);
     const token = window.localStorage.getItem('accessToken');
-    
+
     return (
         <div>
             {token != null ? <NavbarLoggedIn /> : <NavbarNotLoggedIn />}
             <h1>
                 Venue Page under construction ...
             </h1>
-            <SeatMap />
+                <SGStad />
         </div>
 
     )
